+++ test-fixtures/src/bin/async_io.rs
@@ -5,10 +5,10 @@
 use std::io::Write;
 use std::str::FromStr;
 
-use fastly::handle::{BodyHandle, CacheOverride, RequestHandle, ResponseHandle};
-use fastly::http::{HeaderName, HeaderValue, Method, StatusCode, Url};
 use fastly::Error;
 use fastly::Request;
+use fastly::handle::{BodyHandle, CacheOverride, RequestHandle, ResponseHandle};
+use fastly::http::{HeaderName, HeaderValue, Method, StatusCode, Url};
 use fastly_shared::FastlyStatus;
 
 fn is_ready(handle: u32) -> bool {
@@ -103,7 +103,12 @@ fn test_select() -> Result<(), Error> {
     ];
     let mut ready_idx = 0;
     unsafe {
-        let _ = fastly_sys::fastly_async_io::select(handles.as_ptr(), handles.len(), 20, &mut ready_idx);
+        let _ = fastly_sys::fastly_async_io::select(
+            handles.as_ptr(),
+            handles.len(),
+            20,
+            &mut ready_idx,
+        );
     };
     if ready_idx == u32::MAX {
         append_header(&mut ds_resp, "Ready-Index", "timeout");